feat: add --spec to apps deploy
This flag for now will be hidden and will help us transition to a deployment using the intermediate representation.

Current valid values are `latest` or a valid semver. This will be ultimately be validated by Platform API, but this is client validation.

This commit also separates the validation of an app.json and the git environment